# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
  handlers.JAVDbMovieSearchResponse:
    properties:
      code:
        type: string
      cover_url:
        type: string
      detail_url:
        type: string
      rating:
        type: number
      release_date:
        type: string
      title:
        type: string
    type: object
  handlers.ListResponse:
    properties:
      items: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  handlers.LocalMovie:
    properties:
      actress:
        type: string
      code:
        type: string
      fanart_path:
        type: string
      fanart_url:
        type: string
      format:
        type: string
      has_fanart:
        type: boolean
      modified:
        type: string
      path:
        type: string
      size:
        type: integer
      title:
        type: string
    type: object
  handlers.Response:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 智能成人影视库管理系统 - 提供影片管理、种子下载、排行榜等功能
  title: NSFW-Go API
  version: "1.0"
paths:
...
  /local/movies:
    get:
      consumes:
      - application/json
      description: 分页获取本地影片库中的影片列表，支持按女优筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 100
        description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 女优筛选
        in: query
        name: actress
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 影片列表
          schema:
            allOf:
            - $ref: '#/definitions/handlers.ListResponse'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/handlers.LocalMovie'
                  type: array
              type: object
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 获取本地影片列表
      tags:
      - local
  /local/scan:
    post:
      consumes:
      - application/json
      description: 触发一次完整的本地影片库扫描，更新数据库中的影片信息
      produces:
      - application/json
      responses:
        "200":
          description: 扫描成功
          schema:
            $ref: '#/definitions/handlers.Response'
        "500":
          description: 扫描失败
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 手动扫描本地影片库
      tags:
      - local
  /local/stats:
    get:
      consumes:
      - application/json
      description: 获取本地影片库的统计信息，包括总数、女优数量等
      produces:
      - application/json
      responses:
        "200":
          description: 统计信息
          schema:
            $ref: '#/definitions/handlers.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 获取本地影片统计
      tags:
      - local
...
  /search/javdb:
    get:
      consumes:
      - application/json
      description: 在JAVDb网站搜索影片信息，支持番号和标题搜索
      parameters:
      - description: 搜索关键词
        in: query
        name: q
        required: true
        type: string
      - default: movie
        description: 搜索类型
        enum:
        - movie
        - actress
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 搜索结果
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.JAVDbMovieSearchResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: 搜索失败
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: JAVDb在线搜索
      tags:
      - search
  /torrents/search:
    get:
      consumes:
      - application/json
      description: 通过Jackett搜索种子文件，支持任意关键词搜索
      parameters:
      - description: 搜索关键词
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 搜索结果
          schema:
            $ref: '#/definitions/handlers.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: 搜索失败
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 搜索种子文件
      tags:
      - torrents
swagger: "2.0"